initial commit
demonstrating two options for managing site inventory.  the first is
the native ansible way, with site inventories located in the sites/
directory and variables managed with group_vars and host_vars.  the
second is using a site configuration file that is used for managing
all of the site_info variables, as an alternate approach.  playbooks
exist to validate the site inventory for either approach.

any additions to this repo will use the native ansible way of managing
inventory and variables.